Add code for general push send api
@adamgins I haven’t tested this code but you are welcome to test it a
bit.
In short: This code adds a Push.notifications collection so when using
Push.send() its basically an insert to the collection. The
Push.allow/Push.deny:
```js
  Push.allow({
    send: function(userId, notification) {
      // Validate that this user may send this notification
      return true;
    }
  });
```

At the moment the server simply pulls data from the db every 15th sec
or the `sendInterval` given to the Push.init - this ensures that the
server is in control an can throttle messages.

It also gives the sender the ability to see how many recipients the
message had.

I havent implemented the ability to resend the message to new apps
(excluding apps already sent to)

Well 100% untested code - but it gives cliet-side send in a securable
way.

You are welcome to make pr’s to devel if you find bugs here