Fix for issue #181.
Added a try / catch / finally block to push.api.js so that in the Meteor.setInterval function if an error occurs, the isSendingNotification fence will reset if an error is thrown.  This prevents the Meteor process from getting into a state where isSendingNotification is always true and notifications stop until Meteor is reset.